+++ src/glob.ts
@@ -5,6 +5,24 @@ import { conf } from "./conf.js";
 import type { LikeString } from "./with";
 
 /**
+ * Define a custom element.
+ * @param name Name for the new custom element. Must be a valid custom element name.
+ * @param options Object that controls how the element is defined.
+ */
+export const define = (name: string, options?: ElementDefinitionOptions) => (constructor: CustomElementConstructor) => {
+  const tagName = conf.tag(name);
+  if (!tagName) {
+    return;
+  }
+  if (customElements.get(tagName) === undefined) {
+    customElements.define(tagName, constructor, options);
+    conf.enabled.add(name);
+    conf.namemap.set(name, tagName);
+    conf.classmap.set(name, constructor);
+  }
+};
+
+/**
  * Custom CSS variable prefix, join with "--".
  */
 export const cssvar = unsafeCSS("--" + conf.cssvar.replace(/[^a-zA-Z0-9\\-]/g, ""));