+++ src/group/index.ts
@@ -1,2 +1,3 @@
 export * from "./avatar-group.js";
 export * from "./details-group.js";
+export * from "./tab-group.js";