+++ src/items/index.ts
@@ -1,6 +1,7 @@
 export * from "./alert-item.js";
 export * from "./avatar-a.js";
 export * from "./base-button.js";
+export * from "./card-item.js";
 export * from "./dialog-item.js";
 export * from "./link-a.js";
 export * from "./load-track.js";