+++ scripts/dev.js
@@ -1,7 +1,19 @@
+/**
+ * Start the development server.
+ */
+
+import { createServer } from "vite";
 import { cwd } from "node:process";
 import { log } from "node:console";
 import { startDevServer } from "@web/dev-server";
 import { walkExt } from "./lib.js";
+
+/**
+ * Use web/dev-server or vite.
+ *
+ * @type {"web"|"vite"}
+ */
+const server = "web";
 
 (async () => {
   const port = 9527;
@@ -10,14 +22,18 @@ import { walkExt } from "./lib.js";
     log(`http://localhost:${port}/${we}`);
   });
 
-  await startDevServer({
-    config: {
-      nodeResolve: { exportConditions: ["development"] },
-      rootDir: cwd(),
-      port,
-      watch: true,
-    },
-    readCliArgs: false,
-    readFileConfig: false,
-  });
+  if (server === "web") {
+    await startDevServer({
+      config: {
+        nodeResolve: { exportConditions: ["development"] },
+        rootDir: cwd(),
+        port,
+        watch: true,
+      },
+      readCliArgs: false,
+      readFileConfig: false,
+    });
+  } else {
+    await (await createServer()).listen(port);
+  }
 })();