+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Type path.
 
 Move define.
 
+Add assign.
+
 ## 2.3.2
 
 Fix `label-input` styles.